internal

internal documentation

Documentation for QuartetNetworkGoodnessFit's internal functions. Those functions are not exported, but can still be used (like: QuartetNetworkGoodnessFit.foo() for a function named foo()).

index

functions

QuartetNetworkGoodnessFit.dirichlet_max — Method.

dirichlet_max(dcf::DataCF)

Calculate outlier p-values, one for each four-taxon set, using the maximum concordance factor under a Dirichlet distribution. Used by ticr!.

output

vector of outlier p-values, one for each 4-taxon set

value of the concentration parameter α

value of the pseudo likelihood (optimized at α)

source

QuartetNetworkGoodnessFit.dirichlet_min — Method.

dirichlet_min(dcf::DataCF)

First calculate outlier p-values using each of the three concordance factors for each quartet under a Dirichlet distribution, then take the smallest p-value among the three, as the outlier p-value for each four-taxon set.

output

a vector of outlier p-values, one for each quartet

value of the concentration parameter α

value of the pseudo likelihood (optimized at α)

source

QuartetNetworkGoodnessFit.multinom_lrt — Method.

multinom_lrt(dcf::DataCF)
multinom_lrt(quartet::Vector{Quartet})

Calculate the vector of outlier p-values (one per four-taxon set) using the likelihood ratio test under a multinomial distribution for the observed concordance factors.

source

QuartetNetworkGoodnessFit.multinom_pearson — Method.

multinom_pearson(dcf::DataCF)
multinom_pearson(quartet::Vector{Quartet})

Calculate the vector of outlier p-values (one per four-taxon set) using Pearson's chi-squared statistic under a multinomial distribution for the observed concordance factors.

source

QuartetNetworkGoodnessFit.multinom_qlog — Method.

multinom_qlog(dcf::DataCF)
multinom_qlog(quartet::Vector{Quartet})

Calculate the vector of outlier p-values (one per four-taxon set) using the Qlog statistic (Lorenzen, 1995), under a multinomial distribution for the observed concordance factors.

source

QuartetNetworkGoodnessFit.ticr_optimalpha — Method.

ticr_optimalpha(dcf::DataCF)

Find the concentration parameter α by maximizing the pseudo-log-likelihood of observed quartet concordance factors. The model assumes a Dirichlet distribution with mean equal to the expected concordance factors calculated from a phylogenetic network (under ILS). These expected CFs are assumed to be already calculated, and stored in dcf.

When calculating the pseudo-log-likelihood, this function checks the observed concordance factors for any values equal to zero: they cause a problem because the Dirichlet density is 0 at 0 (for concentration α > 1). Those 0.0 observed CF values are re-set to the minimum of:

the minimum of all expected concordance factors, and

the minimum of all nonzero observed concordance factors.

output

maximized pseudo-loglikelihood

value of α where the pseudo-loglikelihood is maximized

return code of the optimization

The optimization uses NLOpt, with the :LN_BOBYQA method. Optional arguments can tune the optimization differently: nloptmethod, xtol_rel (1e-6 by default), starting α value x_start (1.0 by default).

source